+++ src/Input/Input.tsx
@@ -13,14 +13,15 @@ export type InputProps = {
   required?: boolean;
   forwardedRef?: any;
   readOnly?: boolean;
-  mask?: string;
+  mask?: string | string[] | RegExp[];
   onBlur?: React.FocusEventHandler<HTMLInputElement | HTMLTextAreaElement>;
   onFocus?: React.FocusEventHandler<HTMLInputElement | HTMLTextAreaElement>;
   onChange?: React.ChangeEventHandler<HTMLTextAreaElement | HTMLInputElement>;
   onKeyDown?: React.KeyboardEventHandler<HTMLTextAreaElement | HTMLInputElement>;
   onKeyUp?: React.KeyboardEventHandler<HTMLTextAreaElement | HTMLInputElement>;
   disabled?: boolean;
-};
+  type?: string;
+} & Pick<InputMask['props'], 'formatChars' | 'maskChar'>;
 
 function InputBase({
   Component = 'input',